Switched to using publish_parts to get rid of all of the heading information that was getting added to the html using the old publish call.
Updated the default heading level to comply with the moinmoin style. They expect document titles to be h2 and to increment from there. This is my interpretation anyway. Someday need to check to make sure I'm doing this the appropriate way by setting it in my classes constructor. I wasn't sure if this was something I could set in the "settings_overrides" variable passed to publish_parts (not sure what that does).

Fixed a bug where a newline in a nodes astext method was returning a newline in the middle. This was making the moinmoin regexes break. So when calling astext on a node replacing all newlines with nothing. Seems like an appropriate thing to do.


git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk@2624 929543f6-e4f2-0310-98a6-ba3bd3dd1d04