Prevent circular inclusion
Check "include" directives for circular dependencies and raise
a warning (similar to the one rised when file-insertion is disabled).

To prevent endless loops, we
rise a warning when a nested inclusion with the same "path",
"first-line", "end-line", "start-after", and "end-before"
options is detected.
In order to achieve this, the current source's parents are logged.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@8562 929543f6-e4f2-0310-98a6-ba3bd3dd1d04