Write table column widths with 3 digits precision. Fixes bug #444.
Bug #444 reports visible effects of rounding errors with the current
rounding to integer percentages (e.g. 7 columns 13% and one 9% for
8 columns of equal width).

Allow a fractional part (1 digit) when table column widths are
written as percentages in the "colgroup".
Use f-string literal to format coefficient as percentage.

CSS allows decimal numbers (since CSS1) in the specification
of percentages.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9009 929543f6-e4f2-0310-98a6-ba3bd3dd1d04