Make "record_dependecies" test independent of the working dir.
In rST, the location of images and figures is specified as URI
and used literally. As a consequence, finding a corresponding
image file in the local filesystem depends on the current working
directory. The tests are designed to work if the cwd is the test root.
As only files that are actually found are written to the "dependencies",
we disable the tests for image file registry if cwd != test root.

Also, use "pathlib.Path" for simpler path manipulations.

Signed-off-by: milde <milde@users.sf.net>

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9267 929543f6-e4f2-0310-98a6-ba3bd3dd1d04