Use absolute paths in ``test_CLI``
When running tests in working directories other than ``docutils/test``,
the tests fail as the expected files cannot be opened from the
relative ``./data`` path. This change uses the defined ``DATA_ROOT``
constant so that the tests pass regardless of working directory. We
also take the opportunity to explicitly specify the encoding with
which to open the file, avoiding platform-dependent behaviour.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9330 929543f6-e4f2-0310-98a6-ba3bd3dd1d04