New image directive option "loading".
The new option sets the `<image>` element's "loading" attribute
to one of "embed", "link", "lazy" as an indication for the
preferred handling of the image by the writer.

Default and actual handling depend on the writer.
Currently only observed by the HTML5 writer, where it overrides the
"image_loading" configuration setting for the respective image.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9496 929543f6-e4f2-0310-98a6-ba3bd3dd1d04