Fix "latex" writer broken in [r9541].
Adding the "template" setting to the list of registered
"relative paths" prevented finding the default template file
in the LATEX_WRITER_DIR.

Unfortunately, the test suite did not detect this breakage.
It seems only to happen when using the "latex" writer from the command line
(via rst2latex or docutils --writer="latex").

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9581 929543f6-e4f2-0310-98a6-ba3bd3dd1d04