HTML5: use "width" and "height" for unitless image size values.
Fixes [feature-requests:#102]

The HTML `<img>` element supports "width" and "height" attributes to
specify the size or ratio of an image. Values are restricted to
integers (size in pixels).

In Docutils, "width" and "hight" attributes also support a set of CSS
compatible units. Therefore, all size values were passed to HTML via
the common attribute "style".

Using HTML "width" and "hight" attributes for values without unit
provides a means to specify an image's *aspect ratio* without overriding
a size declaration in a CSS style sheet.

This change is not backported to the "html4css1" writer to keep
the output backwards compatible.

The provisional method `HTMLTranslator.image_size()` now returns a
dictionary of attribute values.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9928 929543f6-e4f2-0310-98a6-ba3bd3dd1d04