Revise/fix handling of length specification in the ODT writer.
Use "px" as fallback unit for unitless image size attributes.

Fix conversion factor of "pc" (pica) to "cm".

Fix conversion of image width in "%" if the height is specified.

Adjust fallback DPI value (currently not used) to match CSS units.

Document that PIL/Pillow is always required if images are inserted
without specifying width and height, also if "scale" is not used.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9933 929543f6-e4f2-0310-98a6-ba3bd3dd1d04