Catch deliberate system messages when running the test suite.
When we know a test case triggers a system message,
make sure that it does not clutter the output.

System messages that are only printed to the warning stream may
be checked in an `assertRaisesRegex()` block after setting the "halt_level"
configuration setting to a matching level (cf. e.g. "test_nodes.py"
or the "test/*/test_*_misc.py" test modules).

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@9957 929543f6-e4f2-0310-98a6-ba3bd3dd1d04