LaTeX writer: Don't merge paragraphs if there is a `<target>` between them.
`visit_target()` must keep a blank line between block-level elements.

Most targets with ID are inline elements, because the `PropagateTargets`
transform transfers the ID of a most block-level targets to the next element.
However, a target keeps its ID if the next element is invisible (e.g. a pending
node from a "class" directive, a comment, or a substitution definition).

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@10160 929543f6-e4f2-0310-98a6-ba3bd3dd1d04