Look for valid parent before appending a section element.
Don't append a `<section>` to an `<admonition>` or other body element,
only to <section>` and `<document>`.
Exception: the base node of a nested parser may become the
new section's parent, if it has no parent (detached node).
It is up to the users (or extension implementing a directive) to
prevent or handle these cases.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@10203 929543f6-e4f2-0310-98a6-ba3bd3dd1d04