Deprecate the "match_titles" argument of `states.RSTState.nested_list_parse()`.
`nested_list_parse()` is intended for second and subsequent items of lists
and list-like constructs, it does never match section titles.

The argument value was (mis)used in directive classes to check wheter a
`<topic>` or `<sidebar>` element can be returned.
This check is now replaced by testing for a valid parent node.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@10225 929543f6-e4f2-0310-98a6-ba3bd3dd1d04