Relax test for valid parents of topics or sidebars.
Raise error if the parent element belongs to a not-allowed parent class
instead of when the parent element is not in a supported parent class.
The difference is with uncategorized parent classes (like `node.Element`
or `sphinx.addnodes.only`). This allows for auxiliary wrappers during
the parsing step and fixes issues with the Sphinx "only" directive.

cf. https://github.com/sphinx-doc/sphinx/issues/14002

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@10256 929543f6-e4f2-0310-98a6-ba3bd3dd1d04