More detailled description of the restrictions on transition placement.
The formulation of the restrictions
"a transition may not begin a section or document" and
"a transition may not be at the end of a section or document"
is vague and inconsistent.

Add description of "invisible" elements that are ignored when determining
whether a transition is at the begin or end of a section or the document.

git-svn-id: https://svn.code.sf.net/p/docutils/code/trunk/docutils@10268 929543f6-e4f2-0310-98a6-ba3bd3dd1d04